:adhesive_bandage: Fix: 타입 오류 및 운영 시간 수정

+++ src/main/java/yeoksamstationexit1/recommend/repository/PlaceTimeRepository.java
@@ -15,5 +15,5 @@ public interface PlaceTimeRepository extends JpaRepository<PlaceTime, Integer> {
 
     @Query("SELECT pt FROM PlaceTime pt WHERE pt.place.placeNum = :placeNum " +
             "AND pt.day = :day")
-    Optional<PlaceTime> findByPlaceNumAndDay(@Param("placeNum") Integer placeNum, @Param("day") Integer day);
+    Optional<PlaceTime> findByPlaceNumAndDay(@Param("placeNum") Integer placeNum, @Param("day") Byte day);
 }
